Backend Unit Testing
Implements the backend unit testing:

Authentication: Tests the login and register function.

Market: Tests the ability to Buy and Sell cards, to edit and remove a listing.

Message: Tests the ability to Send and Recieve messages.

Packs and Cards: Tests the ability to open and create packs.

User: Tests the location bonus and Push Token functionallities.